refactor: moved functions to index.js
