+++ functions/index.js
@@ -1,19 +1,51 @@
-/**
- * Import function triggers from their respective submodules:
- *
- * const {onCall} = require("firebase-functions/v2/https");
- * const {onDocumentWritten} = require("firebase-functions/v2/firestore");
- *
- * See a full list of supported triggers at https://firebase.google.com/docs/functions
- */
+const { onDocumentWritten } = require('firebase-functions/v2/firestore');
+const { initializeApp } = require('firebase-admin/app');
+const { getFirestore } = require('firebase-admin/firestore');
 
-const {onRequest} = require("firebase-functions/v2/https");
-const logger = require("firebase-functions/logger");
+// Initialize Firebase Admin SDK
+initializeApp();
+const db = getFirestore();
 
-// Create and deploy your first functions
-// https://firebase.google.com/docs/functions/get-started
+// Recalculate publication counts by year
+async function calculateAggregatedCounts() {
+  const publicationsSnapshot = await db.collection('publications').get();
+  const yearCounts = {};
 
-// exports.helloWorld = onRequest((request, response) => {
-//   logger.info("Hello logs!", {structuredData: true});
-//   response.send("Hello from Firebase!");
-// });
+  publicationsSnapshot.forEach((doc) => {
+    const year = doc.data().year;
+    if (year) {
+      yearCounts[year] = (yearCounts[year] || 0) + 1;
+    }
+  });
+
+  // Convert to required format
+  const aggregatedCounts = Object.entries(yearCounts).map(([label, count]) => ({
+    label,
+    count,
+  }));
+
+  // Save the aggregation
+  await db.collection('aggregations').doc('publicationCounts').set({
+    counts: aggregatedCounts,
+  });
+}
+
+// onDocumentWritten triggers on create, update, or delete
+exports.aggregatePublicationsByYearOnWrite = onDocumentWritten(
+  'publications/{docId}',
+  async (event) => {
+    try {
+      // log before and after data
+      const beforeData = event.data.before.data();
+      const afterData = event.data.after.data();
+
+      console.log(`Document written: ${event.params.docId}`);
+      console.log('Before data:', beforeData);
+      console.log('After data:', afterData);
+
+      await calculateAggregatedCounts();
+    } catch (error) {
+      console.error('Error recalculating counts on document write:', error);
+    }
+  }
+);
